07-organizing the application using specific folders/packages
 - routes
 - middlewares
